requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
six==1.10.0
Werkzeug==0.12.2
